Make bootstrap() work also for matrices
Since the indexing is applied only to the first argument, we cannot use the
ellipsis (...) to pass other vectors that need to be subset. To compute an
AUC we need to index three vectors, which we could pass in as a matrix.

The use of NROW and NCOL allows to extract the correct dimensions for both
vectors and matrices.