important fix for the projection for non Gaussian models:
When we want to project a non gaussian model we solve least squares on a pseudo
gaussian transformation of the expected value of the reference model. This is
currently handled by a surrogate MLE solver such as lme4. Nonetheless, we need
to recompute the approximation at every iteration of the least squares
algorithm. Given that we don't have access to the inner workings of lme4 and we
cannot provide a full custom link function (because we miss the inverse), i have
coded a simple  but ugly surrogate that iterates an MLE solution replacing the
target after every iteration with the new computation. This needs further work
and testing. It seems to work but for instance for a poisson model there are
some bad submodels that I need to look into in more detail.