introduce extract_model_data
This function extracts y, weights and offset from a fit.
For future data we get a newdata argument.
This is passed by the corresponding package (brms, rstanarm), so that
they handle their terms internally, abstracting that functionality from
projpred.

We include a new slot in refmodel and projection objects for
extract_model_data function.

Furthermore, we remove get_refmodel.brmsfit and leverage it to brms
package.