is glm performing iterative least squares?
in theory we fit a gaussian approximation to the non gaussian
observation model by computing pseudo gaussian transformations of the
log likelihood and fitting a Gaussian model instead. Nonetheless, using
lm in the internal iteration returns a much worse approximation than
using glm, but that should be the theoretically correct approach?

The algorithm as I picture it should look like

1. compute pseudo gaussian observation z_i and weights w_i.
2. perform one iteration of weighted least squares with z_i as
target (this approximates a gaussian to the regression coefficients)
3. go back to 1 replacing the current guess with the link prediction of
the previous guess until convergence.