Revert the parallelization of the bootstrap. The first reason is
that in order for the parallel test results to match the sequential ones
exactly, we would have to use doRNG::`%dorng%`() also in the sequential case of
`bootstrap()` and then, for the sequential results in the tests, not register
any **foreach** backend at all or use `registerDoSEQ()`. The second reason is
that, because of the parallelization overhead, the parallelized bootstrap didn't
provide a speed improvement, at least not up to the number of bootstrap samples
used by default (2000).