Docs: Replace "search step" and "evaluation step" by
"search part" and "evaluation part", respectively. The reason is a possible
confusion of the term "search step" with the steps of a stepwise variable
selection, see e.g. `?step`. The second reason is that the term "step" might
suggest an EM-like alternation of steps (which is not the case here, since the
search is completed first and only after that, the evaluation takes place).