Don't use markup in the "title" YAML field because
this causes the following warning when building the **pkgdown** website (see
also r-lib/pkgdown#989); note that setting the option mentioned in the warning
below only helps when building the standalone vignette, not when building the
**pkgdown** website:
```
The vignette title specified in \VignetteIndexEntry{} is different from the
title in the YAML metadata. The former is "projpred: Projection predictive
feature selection", and the latter is "**projpred**: Projection predictive
feature selection". If that is intentional, you may set
options(rmarkdown.html_vignette.check_title = FALSE) to suppress this check.
```