KL divergence aggregation across observations for the `gaussian()` family: Replace
sum by mean (because in `.init_submodel()`, `pseudo_data()$wobs` simply returns
`weights` in that case and those `weights` are not normalized anymore since PR
198). The use of the sum should have had no consequences, though, because in the
currently used KL divergence for the Gaussian family, other constants are
dropped anyway, so dropping 1/N (with N denoting the number of observations) as
another constant shouldn't have done any harm. However, it's probably better to
be explicit and replace the sum by the mean, especially because in the future,
the exact KL divergence might be implemented for the Gaussian family (which
should be achievable by not regarding the dispersion parameter (the residual
standard deviation) as another constant and by not dropping any constants) and
then this aggregation across the observations would already be prepared.

Alternatively, one could have re-introduced the normalization of the observation
weights (i.e., partly reverting PR #198), but for the `binomial()` family,
unnormalized weights indeed make more sense from a theoretical point of view, so
use unnormalized weights for consistency of all families. (The sum of the
observation weights is again a constant, so it shouldn't matter much anyway,
except for the rare case that a user wants to extract the KL divergences without
any constants dropped, but as explained above, for the Gaussian family, other
constants are currently dropped anyway.)