Use `NA` instead of `0` as the "defunct" ("not available") value for predictive variances and dispersion parameters. This
is not only more reasonable from a conceptual point of view, but also allows for
more precise checks (e.g., in the newly introduced check in `.init_submodel()`).
For `datafit`s (but only for those where a dispersion parameter exists), keep
using 0 instead of NA in case of an unspecified `dis`. The reason is that when
using NAs, the NAs propagate down to the results (e.g., `summary()`'s output) and
furthermore, 0 indeed makes sense for `datafit`s because there, we have
`identical(as.vector(<datafit_object>$mu), <datafit_object>$y)` giving `TRUE`.

Note that the change of the default of argument `projpred_var` of
`fit_glm_ridge_callback()` is a no-op change because currently, in all calls to
`fit_glm_ridge_callback()`, argument `projpred_var` is always specified.