Normalize the observation-aggregated KL divergence by
the sum of the observation weights, not by the number of observations. This
makes more sense for the binomial family with > 1 trials (because there,
observations with more trials should probably be given more weight in the
aggregated KL divergence). The fact that in case of the binomial and the Poisson
family, `family$dev.resids()` now receives normalized weights should just be
considered as a technical trick to avoid dividing the S columns of its output by
the normalization constant (and divide the N (usually N << S) observation
weights by the normalization constant instead).

Again, this just affects the normalization constant of the aggregated KL
divergence, so it shouldn't matter much anyway, except for the rare case that a
user wants to extract the KL divergences without any constants dropped, but for
the Gaussian family, other constants are currently dropped anyway.