Remove "make sure correct 'weights' can be found". The reason is that
`weights` is `refmodel$wobs` (see `project_submodel()`) which in turn is a
numeric vector, not a symbol or name or any other language object.

(The help page for `glm()` (and similarly that for `lm()`) states:

> All of weights, subset, offset, etastart and mustart are evaluated in the same
way as variables in formula, that is first in data and then in the environment
of formula.

, so the important point is *evaluation* (of a language object) which is not
necessary here because `weights` is a numeric vector.)