Replace `as.matrix.<class_name>()` methods by
`get_subparams.<class_name>()` (and create a corresponding
generic). The reason is that these methods do not return a
matrix, but a vector. So the naming is kind of confusing.
Furthermore, this is safer in the sense that the
`get_as.matrix_cls_projpred()` workaround for finding all
allowed classes is not necessary (because there is no
default method defined for `get_subparams()`).