Offsets: Wrapper for `ref_predfun` (#250)
* Move the subtraction of the offsets out of the default `ref_predfun` by

introducing a wrapper function around the user-specified (or default)
`ref_predfun`. This has the advantage of avoiding a breaking change for users
with custom `ref_predfun`s.

* Fix a bug when calling `ref_predfun()` with actual `newdata` on

a `stanreg` object with offsets. This was already anticipated in PR #196 and is
now necessary because the subtraction of the offsets was moved out of the
default `ref_predfun` and into a wrapper function. However, this fix is still
just a fragile quick-and-dirty solution. It is more desirable to adopt the
**brms** way, as discussed in PR #196 and issue #218.