A warning for `newdata` not matching the fitted submodel coefficients should be thrown by
the corresponding `predict()` method. All pre-defined `predict()` methods should perform such a check, so only `predict.subfit()` needs to be adapted. The corresponding warning in `proj_helper()` can therefore be removed, which also avoids a definition of a `solution_terms` object which is a bit unsafe (due to the existence of *function* `solution_terms()`). Besides, also move (and modify) a related test from `test_proj_pred.R` to `test_project.R` which is more appropriate there.