At least on my machine, the fix for issue #213 makes RStudio crash when
building the vignette in `R CMD check`. The reason is not clear to me because it does not occur when building the vignette separately. But for now, it's probably better not to show the parallelization in the vignette. The proposed reduction of the submodel fitters' output (see PR #235) might resolve this and then the parallelization could be added back to the vignette.