Allow a custom function (`cvrefbuilder`) for constructing the K reference models in
`.init_kfold_refmodel()`. This is needed for custom reference models, but also
for `get_refmodel.brmsfit()` to be able to set `brms_seed` appropriately there.

Also add element `projpred_k` to the subelements of `cvfits` to be able to infer
`brms_seed`s in `brms::get_refmodel.brmsfit()` from these. This may also be
helpful for custom `cvfits` and/or custom reference models.

Also improve documentation for argument `dis` of `init_refmodel()` because this
new `cvrefbuilder` argument allows to change argument `dis` for the K "refits"
of a `datafit` (if the user actually desires it).