Add a slightly customized `cvrefbuilder` to `get_refmodel.stanreg()` to ensure that
arguments which are passed to `init_refmodel()` by the initial `get_refmodel.stanreg()` call are also passed to `init_refmodel()` by the K `get_refmodel.stanreg()` calls in K-fold CV. This is important for the latent and the augmented-data projection, for example. But it could also be useful for other arguments passed to `init_refmodel()`.