Where possible (non-exported functions which have a calling stack at
the beginning of which a seed is set), don't reset the RNG state on exit. Consequently, at those places, a seed does not even need to be set. However, the latter is just a simplification. The important point is that the RNG state is not reset on exit which avoids that the same seed is re-used multiple times (which is probably bad practice from a theoretical point of view).