`ndraws` and `nclusters` revisited (#291)
* Move `ndraws` default to `nclusters` (because the default of 20 causes

clustering to be used, so it should be less confusing now).

* Merge `NEWS.md` sections for versions unpublished on CRAN into

the section for the most recent version. This should avoid a lot of confusion
and make the `NEWS` file easier to read when it is published on CRAN.

* Update `NEWS.md`, especially with respect to arguments `ndraws`, `nclusters`,

`ndraws_pred`, and `nclusters_pred`.

* Continue "Merge `NEWS.md` sections for versions unpublished on CRAN into

the section for the most recent version. This should avoid a lot of confusion
and make the `NEWS` file easier to read when it is published on CRAN."

Add another important commit which used to be included in v2.0.4.

* Minor fix/improvement in `NEWS.md`.

* `project()`: Never pass `ndraws` to `nclusters` and throw a warning if `ndraws <= 20`.

Thereby, most of the related code is moved to `.get_refdist()` which is also
simplified at multiple places.

* `varsel()` and `cv_varsel()`: Never pass `ndraws` to `nclusters` and

throw a warning if `ndraws <= 20`. Thereby, most of the related code is moved to
`.get_refdist()` and a corresponding change is also necessary for `summary()`
output.

* Adapt the tests. This includes adding option `projpred.warn_ndraws` and

using it to suppress warnings in the tests.

* Re-add `nclusters` to `parse_args_varsel()`.

* For safety: Avoid re-parsing `ndraws` and `nclusters` (and their `_pred` variants) in summary output.

* Fix the decision rule for throwing the warning concerning small `ndraws` and

turn the warning into a message.

* Simplify `print.vselsummary()` (with respect to the number of projected draws).

* Fix `clust_used` in a special case.

* Adapt the tests.

* Add `.get_refdist()` tests.

* Move the definition of `nrefdraws` to `setup.R`.

* `CITATION` file: Add URL for Catalina et al. (2020).

* Update package date.

* Change the order of the clustering cases in `.get_refdist()` to

avoid an inconsistency for `datafit`s, for example (but generally for all
situations where `nclusters == S == 1`). This also allows to remove the special
`S == 1` case in the `nclusters == 1` case.

* `NEWS.md`: Insert PR number.

* Fix an expectation for `datafit`s (discovered while running the

unit tests with `run_more = TRUE`).

* Ensure the test "`cvfits` (actually passed to

init_refmodel()) works for rstanarm reference models" runs through.