Argument `brms_seed` is also needed for reproducible results in
`predict.refmodel()`. In the future, it may be needed at other places, too. And
with the new seed behavior of projpred (see PR #286), K-fold CV might actually
be reproducible even if `brms_seed` is not set (I did not test this, though). So
it's probably best to only vaguely hint at argument `brms_seed` in the vignette.