Always include the intercept in `search_terms`. The reason is
that during the evaluation part of `varsel()` and `cv_varsel()`, the
intercept-only model is always added to the solution path anyway (see `nterms =
c(0, seq_along(search_path$solution_terms))` in the calls to
`.get_submodels()`). Thus, allowing for exclusion of the intercept would have
been harder than forcing its inclusion.

Furthermore, this avoids an issue at the place where `.get_submodels()` indexes
`search_path$submodls` by number (in the case `!refit_prj`) which is dangerous
because of lines
```r
if (is.null(cands))
  next
```
in `search_forward()` which are used in combination with the appending of submodels
(`submodels <- c(submodels, list(subL[[imin]]$submodl))`)---this can lead to
a discrepancy between the index of `search_path$submodls` and the corresponding
model size.