Set `r_eff` to `NA` as discussed with @avehtari. The reasoning is
that in order to calculate the relative efficiencies (using `loo::relative_eff()`), we would have to define an argument such as `chain_id` in `init_refmodel()`, requiring even more input from users in case of a custom reference model. For the PSIS based on the clustered draws (the `validate_search = FALSE` case), calculating the relative efficiencies is probably even less straightforward. The documentation for argument `r_eff` of `loo::psis()` says:

> If r_eff is not provided then the reported PSIS effective sample sizes and Monte Carlo error estimates will be over-optimistic.

so we see that these relative efficiencies are irrelevant for projpred (because projpred does not report PSIS effective sample sizes and neither Monte Carlo error estimates).