Export methods for external generics (for safety and better readability). Also
turn the "pseudo-method" `model.matrix.gamm4()` into an ordinary internal
function. `formula.gamm4()` wasn't a "pseudo-method" (since it was always
applied to objects of class `gamm4`), but exporting it led to an error in the
unit tests, so treat it like a pseudo-method and therefore rename it to
`formula_gamm4()`.