Tests: The avoidance of unnecessary back-and-forth transformations between
latent space and response space seems to have changed the PRNG situation in the
tests to an unfavorable one for `run_more = TRUE`, at least on my machine. As an
ad-hoc solution, change the seed, but as a longer-term objective, we should try
to make the tests more robust against unfavorable PRNG situations.

Some further changes are performed to ensure that the tests can be run with
`run_more = TRUE` again.