The tests comparing `cv_varsel()` results between augmented-data and
traditional projection seem to depend strongly on the seed, so for some test
setups, the tolerance needs to be increased. One issue explained in detail in a
longer in-line code comment does not occur anymore, which could be due to the
fact that unnecessary back-and-forth transformations between latent space and
response space are avoided now (leading to a better numerical accuracy).