Handle the NOTE in `R CMD check` differently (avoid `within()` as
it is probably not intended for use in packages). In principle, we could have
tried `transform()` instead of `within()`, but `?transform` says: "This is a
convenience function intended for use interactively. For programming it is
better to use the standard subsetting arithmetic functions, and in particular
the non-standard evaluation of argument transform can have unanticipated
consequences.".