Require an `S x N` matrix as input for `family$latent_ilink()`, not
an `N x S` matrix. This is done for consistency with other Stan-related R
packages. For brms reference models, this should not have any impact on the
computational efficiency since brms would have to do `t()` internally if not
done in projpred. For rstanarm reference models, this might indeed come with a
worse computational efficiency (because the rstanarm-related code is included in
projpred, so it could perhaps be designed in an efficiency-optimal way), but one
day, projpred might change all of its codebase to the `S x N` convention and
then it would be tedious to require all maintainers of packages providing
`get_refmodel()` methods to change their `family$latent_ilink()` function.
Furthermore, for the augmented-data projection, the same approach was taken (for
the same reasons).