Rename argument `latent_proj` back to `latent` (according to Yann, the
renaming wasn't performed for any crucial reason, so in terms of backwards
compatibility with the original latent projection preprint, it's better to keep
the name `latent` which was used back then).

Furthermore, `init_refmodel()` now doesn't need an own argument `latent` because
it can be passed directly to `extend_family()`.