`brms::categorical()` submodel fitters: Offsets: Add currently
unused code (apart from usage in tests) for recycling a vector of offsets to a matrix with number of columns equal to the number of response categories. This is done for the tests, but also in case a future projpred version performs the submodel fitting with offsets. Note that despite offsets being unexpected in the submodel fitter anyway, it is not clear to me whether `mclogit::mblogit()` supports offsets correctly.