`suggest_size()`: Don't use `subset()` which discards `NA`s in the
logical subsetting condition. Instead, keep rows which have an `NA` in that
logical vector indexing the rows. This is necessary for the latent
projection with `respOrig = FALSE`, `stat %in% c("elpd", "mlpd")`, and
`<refmodel>$dis` consisting only of `NA`s (where `summary.vsel()` returns
only `NA`s for the performance statistic, so the suggested size should also
be `NA`).