Revert "Latent vignette: Switch to (multilevel) Bernoulli model (not finished yet)."
This reverts commit b807ee9c846279326b28a6080228d0a534b3d7ee.

Reason for the revert: For the `binomial()` family, we could now (state: commit
3e47a737504196741ec8f1d5829588bd977aedf2 ("Latent vignette: Activate option
`projpred.PQL` to use `MASS::glmmPQL()` as GLMM submodel fitter.")) successfully
demonstrate that `MASS::glmmPQL()` indeed seems to be preferable over
`lme4::glmer()` because the group-level term is at the same position of the
solution path as in the latent projection (in particular, before the noise
variables). Now, we can try out if `MASS::glmmPQL()` also helps in case of the
`poisson()` family (there, the runtime for the traditional projection was
already an issue).