Disallow latent-scale output for cross-validated variable
selection because the latent response values in `refmodel$y` are the output from
`colMeans(posterior_linpred())` applied to the non-CV'ed reference model, so
using the `fold$omitted` subset of `refmodel$y` as response values in fold k of
K would introduce a dependency between training and test data. The safest way to
achieve this disallowance is to set `refmodel$y` to `NA` (in case of CV).