Submodel predictions: Draw group-level effects randomly for *all* group levels (even existing ones). The
reason is that in a strict sense, the conditional modes of the random effects
(the lme4 "estimates" for the group-level effects; may be termed "BLUPs" under
special circumstances) cannot be regarded as being part of the solution to the
projection problem. With the new behavior, we integrate over the group-level
effects and thereby avoid re-using their "estimates".