`.get_p_clust()` and `.get_refdist()`: Also return "`ilink(link(mu) +(- for some families) offset)`". Use
this in `.init_submodel()` and for `refdist_eval` in `loo_varsel()`.

However, we probably need to tackle the `TODO` comment that was now added to
`.get_p_clust()` (in that case, argument `offs` of `.get_p_clust()` is not
needed anymore and needs to be removed, also in all calls to `.get_p_clust()`)
because we probably need to return the within-cluster aggregated `mu_offs`
instead of "`ilink(link(mu) +(- for some families) offset)`".