`verb_out()`: Replace `message()` by `cat()`, thus writing to
stdout instead of stderr. This is done for consistency with other verbose output
and also because users might want to `sink()` the verbose output to file, but
`sink()`-ing stderr has some pitfalls (see `?sink`). Furthermore, having verbose
output written to stdout puts more emphasis on stderr if there are actually
important messages, warnings, or errors because of the different color that is
used at least in RStudio. Finally, writing verbose output to stdout also seems
to be the preferred way in base R, as can be seen from `txtProgressBar()` which
writes to stdout (by default).