For consistency, use `pct_solution_terms_cv = NULL` not only in
`varsel()` but also in `cv_varsel()` if there is only the full-data search
(i.e., if `validate_search` is `FALSE`). The reason is that users might get a
false impression of the uncertainty in the solution path if `validate_search =
FALSE` is used and `pct_solution_terms_cv` would be an identity matrix (apart
from column `size`), which it used to be in that case until now.