`subfit`s: Fix an error thrown if `newdata` contains only a single
level of a `factor`-coerced `character` variable.

Also fix a bug causing a re-ordering of the levels of a `factor` in
`newdata` (compared to that `factor` in the original dataset) not to get
recognized by `predict.subfit()`, which in turn causes
`x %*% rbind(alpha, beta)` to be incorrect in such cases.

Also fix an error thrown by `fit_glm_ridge_callback()` in case of unused levels
of a `factor`, namely:
```
warning: solve(): system is singular; attempting approx solution
Error: solve(): solution not found
```