Avoid a breaking change for users specifying argument `seed` of
`cv_varsel()` themselves. This concerns K-fold CV based on `cvfun`. The breaking
change consists of the possibility of different results since the PRNG state
when generating the CV folds would be a different one when using the new default
of `seed = NA` in `cv_folds()` (so the folds could be composed differently).

Using `seed = sample.int(.Machine$integer.max, 1)` in this special `cv_folds()`
call should not be a problem as the PRNG state that `cv_folds()` resets to upon
exit is never used again for generating a seed. (After this special `cv_folds()`
call, the next PRNG-using code is `kmeans()`.) Nevertheless, this complicated
construction can be avoided in the feature when introducing other breaking
changes (e.g., in an upcoming UI overhaul).