On Windows, some special handling is needed:
This commit basically reverts the following commits (in the order from top to
bottom):

"Simplify the handling of `...` and `dot_args` as much as possible."
"Further enhance `.noexport`."
"Enhance `.noexport`."
"Add a comment for `.options.snow`."

The reason for the revert of (most of) commit
"Simplify the handling of `...` and `dot_args` as much as possible."
is that on Windows, the 'doParallel' backend complains about incorrect `...`
usage when using the simplified (implicit) `...` code. Instead, the 'doParallel'
backend seems to require `dot_args <- list(...)` to be defined explicitly.

The reason for the revert of commits
"Further enhance `.noexport`."
"Enhance `.noexport`."
(some minor no-op modifications have been performed, too) is that on Windows,
the 'doParallel' package can't find object `refmodel` if it is listed in
`.noexport` (and the same would happen with `y_wobs_test`).

The reason for the revert of commit
"Add a comment for `.options.snow`."
is that I probably misunderstood the implications of `attachExportEnv = TRUE`:
Here, "attaching" is probably the R term as in "attaching a package". So it
shouldn't affect the runtime (at least not heavily).

Also note that on Windows, the 'doParallel' backend would allow `.export` to be
omitted, but the 'doFuture' backend requires `.export` to be specified (at least
when using the 'callr' plan for package 'future').