Fix a bug for the augmented-data projection in combination with subsampled LOO:
When subsetting the rows of an augmented-rows matrix or the elements of an
augmented-length vector, attribute `nobs_orig` was not adapted.

Note that subsetting the rows of an augmented-rows matrix is only legal in terms
of the observations (individuals), not in terms of the (possibly latent)
response categories.

Similarly, subsetting an augmented-length vector is only legal in terms of the
observations (individuals), not in terms of the (possibly latent) response
categories.