`get_refmodel.default()`: It is redundant to require `object` to have
an element `data` which contains the same dataset as the user needs to supply to
`init_refmodel()`'s argument `data` anyway. Furthermore, it is quite restrictive
to assume `object` to be a `list`.

Furthermore, it is better to check for non-`NULL` `newdata` in
`.extract_model_data()` on a case-by-case basis.