Create `p_sel` inside of `select()` and `p_pred` inside of `get_submodls()`
(without further adjustments, then in `cv_varsel()`, the PRNG state for the
full-data search's `p_sel` would be affected by this change if `validate_search`
is `TRUE`, so this is avoided by moving the `select()` call for the full-data
search upwards).

This change should also provide an efficiency improvement because in case of a
large number of observations, the `get_refdist()` output can be large (and with
this change, this large output is stored only as long as necessary).