Make rstanarm reference models behave more similarly to brms reference models
concerning observation weights and offsets.

Also wrap the user-specified `extract_model_data` function in an
`init_refmodel()`-internal `extract_model_data` function in order to retrieve
the correct `newdata` when `newdata` is `NULL` (originally, I thought this was
important for K-fold CV, but I don't think so anymore; however, it is beneficial
already from a user perspective because a special handling of `newdata = NULL`
in the user-specified `extract_model_data` function is not necessary anymore).