`summary.vsel()`: Rename column `solution_terms` to `ranking_fulldata`
and also remove the message added in an earlier release (where `ranking()` was
introduced, I think), stating that column `solution_terms` contains the
full-data predictor ranking.

This is mainly in anticipation of the upcoming UI overhaul (and in line with the
recent deprecation of the `solution_terms()` function), but can be performed
already now because we are making some larger changes regarding element
`selection` of `vselsummary` objects anyway.