Fix a bug when bootstrapping: In the case `if (!is.null(mu.bs))`, we
need to call `bootstrap()` only once (making use of the matrix ability of
`bootstrap()`'s `fun`) instead of twice, to avoid that the bootstrap samples
differ between the submodel (`mu`) and the baseline model (`mu.bs`). The old
behavior led to the strange result that the reference model's SE could be
non-zero in case of `deltas = TRUE` when a seed was not set via argument `seed`
(which gets passed to `bootstrap()`).