Fix a bug in `get_stat()`'s handling of `NA`s in case of
subsampled PSIS-LOO CV (`nloo`) with `deltas = TRUE` and `baseline = "best"`.

For `baseline = "ref"`, this is only a refactor improving the safety and
readability of `get_stat()`'s handling of `NA`s because for `baseline = "ref"`,
we should always have no `NA`s in `lppd.bs` and `mu.bs`, so in that case,
`n_notna` did not require an adjustment and also because the math operations
connecting `mu` and `mu.bs` (analogously for `lppd` and `lppd.bs`) ensured that
only the "inner join" of non-`NA` elements (i.e., the set of observations for
which both `mu` and `mu.bs` (analogously for `lppd` and `lppd.bs`) are not `NA`)
is used.

This addresses question 1 from
<https://github.com/stan-dev/projpred/issues/94#issuecomment-1246694046>.