Add argument `search_control` accepting tuning parameters
(in R also known as "control arguments") for the search.

The main reason for this is to allow different tuning parameters for
`refmodel$div_minimizer()` in `search_forward()` and `perf_eval()`. However,
this change is also the most straightforward solution to ensure that fold-wise
searches from `cv_varsel.vsel()` use the same tuning parameters as the
previously run full-data search (`args_search` did not take `...` into account).