Need to take a different approach for avoiding further warnings
in newer brms versions because the previous approach was problematic for
rstanarm reference models (sometimes, an error was thrown because the internally
added data columns of observation weights and/or offsets could not be found in
`newdata`).

Part 1: Move the `extract_model_data()` call from `proj_helper()` into the
`onesub_fun()`s.