Revert "UNFINISHED: move out the new "mixed deltas" variant of `plot.vsel()`, the"
This reverts commit ea18e9aef11d61c7d3851510ef5a13f8ba7e7047.

Reason for the revert: It's easier if we first add everything that should be in
the final (separate) pull request that incorporates all changes from this pull
request (including the new "mixed deltas" plot, the new column behavior of
`summary.vsel()` etc.). Afterwards, we can simplify the current pull request and
leave things like the new "mixed deltas" plot and the new column behavior of
`summary.vsel()` for the separate pull request.