Since `summaries_fast` is created by a call to `loo_varsel()` with
`validate_search = FALSE`, the search is not run again when creating
`summaries_fast`. Only the performance evaluation (including the re-projections
required for it) is run again. Hence, it would be inconsistent to treat
`summaries_fast` like the search-related arguments of `cv_varsel.refmodel()`
when calling `cv_varsel.refmodel()` from within `cv_varsel.vsel()`. Thus, a lot
of code related to `summaries_fast` can be removed, which is done here.