latent-projection vignette: perform several enhancements in the
censored-observations examples

First, we change the censoring times to avoid
problems in the log-normal example (there, it seems we had too many censored
observations).

Second, we change the true `sdlog` parameter to avoid
problems in the log-normal example (there, it seems we had a too right-skewed
distribution).

Third, we remove the `set.seed()` call because we already have one at the
beginning of the vignette.